+++ packages/tolk/lib/uop/tolk_uop.ml
@@ -11,6 +11,7 @@ module Dtype = Dtype
 module Const = Const
 module Bound = Bound
 module Target = Target
+module Tiny_elf = Tiny_elf
 module Uop = Uop
 module Render = Render
 module Upat = Upat