+++ packages/tolk/test/unit/engine/test_jit.ml
@@ -55,7 +55,7 @@ let allocator =
 let make_device ?(name = "TEST:0") ?(state = runtime_state ()) () =
   Device.make ~name ~allocator
     ~renderer_set:(Device.Renderer_set.make ~device:"TEST" [ "TEST", Fun.const renderer ])
-    ~runtime:(fun _ _ ->
+    ~runtime:(fun _ ->
       {
         Device.call =
           (fun _ ~global ~local:_ ~vals ~wait:_ ~timeout:_ ->
@@ -73,7 +73,7 @@ let device = make_device ()
 
 let to_program body =
   let info = U.program_info_from_sink body in
-  U.program ~sink:body ~linear:(U.linear []) ~source:(U.source "")
+  U.program ~sink:body ~linear:(U.linear (U.toposort body)) ~source:(U.source "")
     ~binary:(U.binary "") ~info ()
 
 let shape_const n = U.const (Const.int Dtype.weakint n)
@@ -148,7 +148,7 @@ let make_kernel_jit ?(body = U.sink ~kernel_info:(kernel_info "jit_k") [])
 let graph_device ~max_buffer_offset =
   Device.make ~name:"TEST:0" ~allocator
     ~renderer_set:(Device.Renderer_set.make ~device:"TEST" [ "TEST", Fun.const renderer ])
-    ~runtime:(fun _ _ -> failwith "unused")
+    ~runtime:(fun _ -> failwith "unused")
     ~synchronize:(fun () -> ())
     ~graph:
       {